+++ rust/src/core.rs
@@ -1,0 +1,8 @@
+// Copyright 2023 ReductStore
+// This Source Code Form is subject to the terms of the Mozilla Public
+//    License, v. 2.0. If a copy of the MPL was not distributed with this
+//    file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
+pub mod env;
+pub mod logger;
+pub mod status;