+++ src/reduct/api/console.h
@@ -7,14 +7,14 @@
 #define REDUCT_STORAGE_CONSOLE_H
 
 #include "reduct/api/common.h"
-#include "reduct/asset/asset_manager.h"
+#include "rust/rust_part.h"
 
 namespace reduct::api {
 
 class Console {
  public:
-  static core::Result<HttpRequestReceiver> UiRequest(const asset::IAssetManager* console, std::string_view base_path,
-                                              std::string_view path);
+  static core::Result<HttpRequestReceiver> UiRequest(const rust::Box<rust_part::ZipAssetManager>& console,
+                                                     std::string_view base_path, std::string_view path);
 };
 
 }  // namespace reduct::api