+++ src/reduct/api/server_api.cc
@@ -5,12 +5,8 @@
 
 #include "reduct/api/server_api.h"
 
-#include "reduct/auth/token_auth.h"
-
 namespace reduct::api {
 
-using auth::ITokenRepository;
-using auth::ParseBearerToken;
 using core::Error;
 using core::Result;
 using storage::IStorage;
@@ -21,14 +17,12 @@ Result<HttpRequestReceiver> ServerApi::Info(const IStorage* storage) { return Se
 
 Result<HttpRequestReceiver> ServerApi::List(const IStorage* storage) { return SendJson(storage->GetList()); }
 
-core::Result<HttpRequestReceiver> ServerApi::Me(const auth::ITokenRepository* token_repo,
+core::Result<HttpRequestReceiver> ServerApi::Me(const rust_part::TokenRepository& token_repo,
                                                 std::string_view auth_header) {
-  auto [token_value, error] = ParseBearerToken(auth_header);
-  if (error) {
-    return error;
-  }
+  auto token = rust_part::new_token();
+  auto err = rust_part::token_repo_validate_token(token_repo, std::string(auth_header), *token);
 
-  return SendJson(token_repo->ValidateToken(token_value));
+  return SendJson(Result{rust_part::token_to_json(*token), Error::FromRust(*err)});
 }
 
 }  // namespace reduct::api